Review comments

+++ drake/automotive/test/idm_planner_test.cc
@@ -23,9 +23,9 @@ class IdmPlannerTest : public ::testing::Test {
   double get_v_0() const { return v_0_; }
 
   void SetInputValue(const std::vector<double>& state) {
-    unsigned int state_size =
+    int state_size =
         dut_->get_ego_port().get_size() + dut_->get_ego_port().get_size();
-    DRAKE_DEMAND(state_size == state.size());
+    DRAKE_DEMAND(state_size == (int) state.size());
     // Get the state values.
     const double x_ego = state[0];
     const double v_ego = state[1];
